Refactor: record extension recommendation events (#65)
* Refactor: record extension recommendation events

* Fix: initialize utils before use
