+++ src/overview/assets/index.html
@@ -78,6 +78,7 @@
             </div>
             <div ext="redhat.vscode-xml">
               <a href="command:java.helper.showExtension?%22redhat.vscode-xml%22" title="Schema validation when editing pom.xml...">Install XML extension</a>
+            </div>
             <div ext="vscjava.vscode-java-dependency">
               <a href="command:java.helper.showExtension?%22vscjava.vscode-java-dependency%22" title="Install Java Dependency Viewer extension...">Install Java Dependency Viewer</a>
             </div>
